claude_llama/requirements.txt
llama-cpp-python>=0.2.0
sounddevice>=0.4.6
numpy>=1.24.0
piper-tts>=1.2.0

# Optional: for GPU acceleration, see README.md for CMAKE_ARGS